pyproject.toml
[tool.poetry]
name = "admin_setup"
version = "0.1.0"
description = ""
authors = ["Shubham Biswas <shubhambiswas.zero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.2"
databases = {extras = ["postgresql"], version = "^0.3.2"}
gino = "^1.0.0"
fastapi = "^0.54.2"
python-dotenv = "^0.13.0"
pyjwt = "^1.7.1"
permissions-system = "^0.1.1"
uvicorn = "^0.11.5"
passlib = "^1.7.2"
python-multipart = "^0.0.5"
psycopg2-binary = "^2.8.5"
email_validator = "^1.1.1"
bcrypt = "^3.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
